Cycle through all available jobs until match found.
Previously, only the first job in the queue was scanned to see if a
worker was available.  This means that jobs were only executed on a
strict first-come-first-served basis.  Now all queued jobs are
checked each cycle, and when the first match is found that job is
run and the rest of the network cycle continues.